Fixed typo of the word successfully. Made it build the multipart body for http using a category to save the amount of code needed to type to add a parameter or file. Added an option to change the length of time to limit files to when uploading to allow people to keep files in the folder of which the auto upload filter is set to, by default the time length is 5 seconds. Fixed a possible crash where it may not beable to get the value of a metadata. Updated the front window system to work with carbon events to improve the acuracy of which application is front and also count the amount of windows that were opened to prevent it from going to the previous front application when there are still windows open. Fixed some leaks and fixed some possible logic errors. Made CocoaShare localizable. Added translation for Portuguese. Updated GeckoReporter to 0.2.

+++ Classes/twitpic/MGMTwitpicPostWindow.m
@@ -10,8 +10,8 @@
 #import "MGMAddons.h"
 
 @implementation MGMTwitpicPostWindow
-- (id)initWithContentRect:(NSRect)theRect styleMask:(unsigned int)theStyleMask backing:(NSBackingStoreType)theBufferingType defer:(BOOL)isDefer {
-	if (self = [super initWithContentRect:theRect styleMask:NSBorderlessWindowMask backing:theBufferingType defer:isDefer]) {
+- (id)initWithContentRect:(NSRect)theRect styleMask:(NSUInteger)theStyleMask backing:(NSBackingStoreType)theBufferingType defer:(BOOL)isDefer {
+	if ((self = [super initWithContentRect:theRect styleMask:NSBorderlessWindowMask backing:theBufferingType defer:isDefer])) {
 		[self setLevel:NSStatusWindowLevel];
         [self setAlphaValue:1.0];
         [self setOpaque:NO];
